Introducing Role-Based Access Control
That's it. Now there's a system that controls who's able to do what.
Particularly:

- There's an admin (only one, with username: nimmos / password:
12345678). This admin can post, make a comment and edit any post in the
blog.
- Logged in users can post, make a comment and edit only their own
posts. Every new user has these features automatically assigned to them
when they sign up to the blog.

To make it work like that, I did the following:

- I made a migration and some changes to the application configuration
to use RBAC (Role-Based Access Control).
- RbacController creates the permissions, roles and the assignments. It
also assign the 'admin' role to the user with ID=1, which is the first
user in the tbl_user.
- 'Edit post' or 'Post' buttons appear or disappear depending on the
type of user.

Another changes:

- Now everything is done with the models TblUser, TblPost and
TblComment.
- Minor changes in signup view controls distribution. I wanted to play
with offsets and such. (See it in large screens)
- I think everything looks stylishly good, for now.