Filemanager does its thing while I play with Slick
Yup, these are some of the good changes I've made:

- Fiddled into the folder organization of /uploads/, now it should work
perfectly with regular upload (for the post header image) and Responsive
Filemanager.
- Responsive Filemanager: now it selects the root folder depending on
whether you're an admin or just an author. It uses:
$_SESSION["RF"]["subfolder"] = ""
- Slick: I've been playing with it. Still doesn't have a real function
but it looks neat. I think I'm doing some progress.

Now for the additional changes:

- Post comment section is now a ListView.
- In case .gitkeep isn't working, I'm uploading a third-party.rar with
the needed elements of the vendor folder. Just unzip it in the
corresponding folder without touching nothing else.