Introducing UserController and different layouts
Check out this new controller: UserController. It will let the
authenticated users configure their profile info. Moreover:

- Different layouts for different types of users!!! Just "main.php" (for
guest users) and "user.php" (for admins and authenticated) for now.
- Post and User controllers check the user's role at the beginning of
the controller's initialization ( init() ) and store this info in a
$_SESSION variable.
- "profile.php" isn't finished yet. Just testing if it displays the
profile image.

Other things that the blog does (that it didn't do before!):

- It saves profile images. Its input is at the signup page.
- When a post is removed, its images (and thumbnails) are also removed.
I'm thinking in providing a checkbox to let the user decide if the
images stay or not.
- Changed the image folders distribution. Now is something like this:

1. rfmimages: images upload folder for Responsive Filemanager.
2. rfmthumbs: thumbnails for Responsive Filemanager.
3. uploads: contains "post" for post header images and "user" for user
profile images.