Accepting the cookies policy
- Added the message for "accepting cookies". It consists on a floating
message at the top of the page (under the navbar) that, once accepted,
it sends a cookie to the user.
- Layouts: I've renamed the layouts to "user" and "guest". Separated the
content to a file called "content.php". This file is the one that
contains the "cookie accepting" message.
- SiteController now has role check for assigning layout.
- Posts have their author's profile thumbnail in the author's data.
- Comments have their author's profile thumbnail. Rearranged the
elements in the comment view.
- Improved the index post-list style. I've also changed some minor
things for each item of that list.
- It shows the post total count in the admin profile page. Not visible
for other users.