Tags assignment complete
A post is able to have its own tags! No post is free from the
possibility of being tagged.

Tags:

- Now every post display its tags.
- The user can edit the post tags. After editing, if a tag doesn't
exist, it gets created, and if a tag is stopped being used by any post,
it gets deleted.

Security:

- I've set the beforeAction() method to permit access to certain actions
depending on the type of user. For security, it also checks if the user
is admin or author of the post when editing or deleting.

Minor changes:

- In post-compose I've prevented the ENTER KEY from submitting the form.
It was interfering with bootstrap tag input.
- Deleted some unnecesary things.